+++ src/pages/footer.js
